+++ src/baobab_activity_reporting/storage/__init__.py
@@ -1,0 +1,4 @@
+"""Sous-package de la couche de persistance.
+
+Contient le gestionnaire de session SQLite et les repositories.
+"""